ZEPPELIN-3617. Allow to specify saving resourceName as paragraph property
### What is this PR for?
This is allow user to specify resourceName when they want to save the paragraph result into ResourcePool. Before this PR, user don't have control on what name of the saving resource name. It is associated with noteId and paragraphId, but this is not a good solution. Because when you clone the note, noteId will be changed, and you have to change the noteId in code as well. This PR is trying to allow user to set resource Name for the saving paragraph result.

### What type of PR is it?
[ Feature |g]

### Todos
* [ ] - Task

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-3617

### How should this be tested?
* CI pass

### Screenshots (if appropriate)
![screen shot 2018-07-12 at 2 45 54 pm](https://user-images.githubusercontent.com/164491/42616882-5ee2fa3a-85e2-11e8-80d5-3ba45c84114b.png)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Jeff Zhang <zjffdu@apache.org>

Closes #3065 from zjffdu/ZEPPELIN-3617 and squashes the following commits:

024965e64 [Jeff Zhang] ZEPPELIN-3617. Allow to specify saving resourceName as paragraph property