[ZEPPELIN-3610] Cluster Raft module design
What is this PR for?

By using the Raft protocol, multiple Zeppelin-Server groups are built into a Zeppelin cluster, the cluster State Machine is maintained through the Raft protocol, and the services in the cluster are agreed upon. The Zeppelin-Server and Zeppelin-Interperter services and processes are stored in the Cluster MetaData. Metadata information;

What type of PR is it?

[Feature]

Todos

- - Added support for checking lambda syntax styles in the pom.xml file
  - add raft algorithm copycat jar
  - add  Cluster State Machine
  - add state machine query command
  - add state machine delete command
  - add state machine put command

What is the Jira issue?

- https://issues.apache.org/jira/browse/ZEPPELIN-3610

How should this be tested?

CI pass

Screenshots (if appropriate)

Questions:

- Does the licenses files need update? No
- Is there breaking changes for older versions? No
- Does this needs documentation? No