[ZEPPELIN-3610] Cluster Raft module design
By using the Raft protocol, multiple Zeppelin-Server groups are built into a Zeppelin cluster, the cluster State Machine is maintained through the Raft protocol, and the services in the cluster are agreed upon. The Zeppelin-Server and Zeppelin-Interperter services and processes are stored in the Cluster MetaData. Metadata information;

[Feature]

* [x] added support for checking lambda syntax styles in the pom.xml file
* [x] add raft algorithm copycat jar
* [x] add Cluster State Machine
* [x] add state machine query command
* [x] add state machine delete command
* [x] add state machine put command

* https://issues.apache.org/jira/browse/ZEPPELIN-3610

CI pass

* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? Yes