[ZEPPELIN-3656] Fix for completion with Livy interpreter
### What is this PR for?
Fix for NullPointerException when using code completion in the Livy Interpreter when Shared Interpreter is enabled.

### What type of PR is it?
Bug Fix

### What is the Jira issue?
[ZEPPELIN-3656](https://issues.apache.org/jira/browse/ZEPPELIN-3656)

### How should this be tested?
Run Livy Interpreter in an environment where Shared Interpreter is enabled and attempt to trigger code completions.

A unit test has been added to test the completion API when using a shared interpreter. Before applying the fix this test was failing with the same NPE.

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Alex Bush <alex.bush@coxauto.co.uk>

Closes #3094 from alexjbush/ZEPPELIN-3656 and squashes the following commits:

4f4a875bf [Alex Bush] Correct completion test
cd0c2133f [Alex Bush] Force a failing test
71562e0f1 [Alex Bush] Test IT for Changes
32aa7c567 [Alex Bush] [ZEPPELIN-3656] Fix for completion with Livy interpreter