ZEPPELIN-3699. Remove the logic of converting single result to multiple results
### What is this PR for?
It is a refactoring PR. Zeppelin introduce multiple interpreter result in 0.7, and also introduce logic of converting single result to multiple results. I think it it good time to remove that for 0.9, as I believe no one is using zeppelin before 0.7 for now.

### What type of PR is it?
[Refactoring]

### Todos
* [ ] - Task

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-3699

### How should this be tested?
* CI pass

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Jeff Zhang <zjffdu@apache.org>

Closes #3136 from zjffdu/ZEPPELIN-3699 and squashes the following commits:

e6298c786 [Jeff Zhang] ZEPPELIN-3699. Remove the logic of converting single result to multiple results