ZEPPELIN-3858. Code completion doesn't work for spark interpreter when cursor is in the middle of code
### What is this PR for?
It is a trivial fix for the code completion of spark interpreter when the cursor is in the middle of code.

### What type of PR is it?
[Bug Fix]

### Todos
* [ ] - Task

### What is the Jira issue?
* https://jira.apache.org/jira/browse/ZEPPELIN-3858

### How should this be tested?
* CI pass

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Jeff Zhang <zjffdu@apache.org>

Closes #3221 from zjffdu/ZEPPELIN-3858 and squashes the following commits:

bb56b723e [Jeff Zhang] ZEPPELIN-3858. Code completion doesn't work for spark interpreter when cursor is in the middle of code

(cherry picked from commit de17d2b8758d9ce83beb1e3e54645de4d831ea1f)
Signed-off-by: Jeff Zhang <zjffdu@apache.org>