[ZEPPELIN-3511] remove old button "Download Data as CSV/TSV"
### What is this PR for?
This PR removes the legacy method (button and dropdown near the top left of each paragraph) for exporting the displayed data as a CSV or TSV file. The widgets of the legacy method are seen in the images below.

![](https://user-images.githubusercontent.com/477015/41198653-71eb5646-6ca0-11e8-9e05-988d6b9a58e5.png)
![](https://user-images.githubusercontent.com/477015/41198654-7223e7d6-6ca0-11e8-8f4d-73f4c8e9234a.png)

The legacy method is now redundant as a new improved data-export menu is available (top right corner of data display grid). The new method also does not suffer from certain issues in the legacy method. The new method is seen in the image below.


![](https://user-images.githubusercontent.com/477015/41198662-c527c7ea-6ca0-11e8-9a29-e17b92392d5c.png)


### What type of PR is it?
[Improvement]

### What is the Jira issue?
https://github.com/TinkoffCreditSystems/zeppelin/pull/15
https://issues.apache.org/jira/browse/ZEPPELIN-3511
https://github.com/apache/zeppelin/pull/3013
ZP-19

### How should this be tested?
1. Travis CI pass
2. Manual checking (see screenshot below)

### Screenshots (if appropriate)
![](https://user-images.githubusercontent.com/477015/41198641-47f2777a-6ca0-11e8-8289-dbc0978468ec.png)

### Questions:
* Does the licenses files need update? - No
* Is there breaking changes for older versions? - No
* Does this needs documentation? - No

cabb83e [Sanjay Dasgupta] z3511-remove-old-download-button: Initial updates