[ZEPPELIN-3898] Adding repo in testDelRepo so test can run by itself
### What is this PR for?
Test DependencyResolverTest.testDelRepo fails when run on its own. The test needs a repo to already be added so it can delete it. If the test is run on its own, the repo is not added to start and so the test fails. The proposed PR fixes this issue by having the test add the repo at the beginning.

### What type of PR is it?
Bug Fix

### What is the Jira issue?
https://issues.apache.org/jira/browse/ZEPPELIN-3898

Author: OrDTesters <ordtesters@gmail.com>

Closes #3249 from OrDTesters/fix-DependencyResolverTest-testDelRepo and squashes the following commits:

13a282252 [OrDTesters] [ZEPPELIN-3898] Adding repo in testDelRepo so test can run by itself