[ZEPPELIN-3919] Paragraph config default value can be customized
The title of the paragraph of the current notebook is not displayed by default, and the runOnSelectionChange check option is ON by default.
For added flexibility, add the config configuration to the `interpreter-setting.json` file in the interpreter. You can modify the default values with the following four options.

```
"config": {
  "fontSize": 9,
  "colWidth": 12,
  "runOnSelectionChange": false,
  "title": true
}
```

In the `hadoop submarine interpreter`, each paragraph of a notebook represents a complete module of a machine learning algorithm. The name of the algorithm name needs to be filled in through the title of the paragraph.

**Notebook paragraph display title (Optional)**
The notebook paragraph does not display the title by default.
You can have the title of the notebook display the title by `config.title=true`.

**Notebook run on selection change (Optional)**
The dynamic form in the notebook triggers execution when the selection is modified.
You can make the dynamic form in the notebook not trigger execution after selecting the modification by setting `config.runOnSelectionChange=false`.

[Improvement]

* [x] Add paragraph config default configuration in interpreter-setting.json

* https://issues.apache.org/jira/browse/ZEPPELIN-3919

CI pass

![alt text](https://github.com/liuxunorg/images/blob/master/zeppelin/config-setting.gif?raw=true "config-setting")

* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? Yes