[ZEPPELIN-3920] Paragraph support custom menu display with multiple URL links
### What is this PR for?

Now the execution of the paragraph in zeppelin only supports the display of a single URL link, for example: Spark paragraph execution is a Spark JOB URL link.
The machine learning task in the hadoop submarine needs to provide the user with a tensorflow Parameter server, a tensorflow worker server log link in YARN, and a tensorboard link, so a custom menu display that supports multiple URL links is now available.

### What type of PR is it?
[Improvement]

### Todos
* [x] Add testRuntimeInfos test cases in NotebookServerTest.java
* [x] modify paragraph-control.html

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-3920

### How should this be tested?
CI pass

### Screenshots (if appropriate)

![alt text](https://github.com/liuxunorg/images/blob/master/zeppelin/multi-job-url.gif?raw=true "multi-job-url.gif")

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No