[ZEPPELIN-3919] Paragraph config default value can be customized
### What is this PR for?

The title of the paragraph is not displayed by default, and the runOnSelectionChange check option is `ON` by default.
For added flexibility, Add the config configuration section in the interpreter-setting.json file of the interpreter. You can modify the default values with the following four options.

```
"config": {
  "fontSize": 9,
  "colWidth": 12,
  "runOnSelectionChange": false,
  "title": true
}
```

In the `hadoop submarine interpreter`, each paragraph of a notebook represents a complete module of a machine learning algorithm. The name of the algorithm name needs to be filled in through the title of the paragraph.

**Notebook paragraph display title (Optional)**
The notebook paragraph does not display the title by default.
You can display the title by default with `config.title = true`.

**Notebook run on selection change (Optional)**
The dynamic form in the notebook triggers execution when the selection is change.
You can make the dynamic form in the notebook not trigger execution after selecting the modification by setting `config.runOnSelectionChange=false`.

### What type of PR is it?
[Improvement]

### Todos
* [x] Add paragraph config default configuration in interpreter-setting.json

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-3919

### How should this be tested?
[CI pass](https://travis-ci.org/liuxunorg/zeppelin/builds/470852962)

### Screenshots (if appropriate)

![alt text](https://github.com/liuxunorg/images/blob/master/zeppelin/config-setting.gif?raw=true "config-setting")

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? Yes

Author: liuxunorg <33611720@qq.com>

Closes #3271 from liuxunorg/ZEPPELIN-3919 and squashes the following commits:

58c0f4451 [liuxunorg] Add test case & Optimization code.
6dfb8a15f [liuxunorg] Paragraph configuration name using static variables
0963a89ab [liuxunorg] Support for dynamic modification of interpreter types
5dbb559ad [liuxunorg] [ZEPPELIN-3919] Paragraph config default value can be customized