[ZEPPELIN-3931] Redisplay angularObjectBind when the note is reopened
### What is this PR for?
At present, Bind's angularObject in note is only valid in the current operation web page.
When the note is reopened, or the zeppelin service is restarted, the angularObject of Bind in the note cannot be displayed, and the bind operation must be repeated.

The submarine has a lot of startup commands and parameters. In order to provide the best experience for the user, we provide the parameters to the user through the WEB control.

Zeppelin's own dynamic form is more suitable for parameter input in the query, but it does not meet the needs of the submarine interpreter, so the submarine interpreter uses the angular template to generate a richer input interface.
The controls in the interface are saved to the note through Bind's angularObject, so that when the user reopens the note, there is no need to re-enter it.

### What type of PR is it?
[Improvement]

### Todos
* [x] Save angularObject to Note in NotebookServer::angularObjectClientBind(...) function.
* [x] Delete angularObject to Note in NotebookServer::angularObjectClientUnbind(...) function.
* [x] Update angularObject to Note in NotebookServer:: angularObjectUpdated(...) function.
* [x] Load angularObject update to Note in NotebookServer::getNote(...) function.
* [x] Add test case in NotebookServerTest::testAngularObjectSaveToNote(...).
* [x] Add test case in NotebookServerTest::testLoadAngularObjectFromNote(...).

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-3931

### How should this be tested?
[CI pass](https://travis-ci.org/liuxunorg/zeppelin/builds/473897099)

### Screenshots (if appropriate)

![alt text](https://github.com/liuxunorg/images/blob/master/zeppelin/angularBing-save.gif?raw=true "angularBing-save.gif")

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: liuxunorg <33611720@qq.com>

Closes #3278 from liuxunorg/ZEPPELIN-3931 and squashes the following commits:

96cd90b44 [liuxunorg] ### What is this PR for?