[ZEPPELIN-4004] Fix RemoteResource.invokeMethod()
### What is this PR for?
RemoteResource is a representation of object in ResourcePool in Interpreter running on another process.

RemoteResource provides a invokeMethod() to call method of an object remotely, which is not working now. This PR provide a fix for that. Also improve the test case to cover this case.

Also provide small improvement to inference parameter type on remote invocation.

### What type of PR is it?
Bug Fix, Improvement

### What is the Jira issue?
https://issues.apache.org/jira/browse/ZEPPELIN-4004

### How should this be tested?
Unittest included. CI pass.

### Questions:
* Does the licenses files need update? no
* Is there breaking changes for older versions? no
* Does this needs documentation? no

Author: Lee moon soo <moon@apache.org>

Closes #3312 from Leemoonsoo/ZEPPELIN-4004 and squashes the following commits:

c44dcc073 [Lee moon soo] address reference to invokeMethod is ambiguous
1893ad150 [Lee moon soo] inference parameter type
fe6f5ed83 [Lee moon soo] Fix RemoteResource.invokeMethod() and update testcase to cover it