[ZEPPELIN-4019] Interpreter process get the note from zeppelin server
The interpreter is executed according to the paragraph. Click the run button in the paragraph, and only the code content in the current paragraph can be obtained in the interpreter process.

To pass the Dashboard paragraph in the submarine interpreter, you need to submit all the code content in this note to tensorflow for execution.

So, I think need to get the note data in sever through `RemoteInterpreterEventServer` in the interpreter process.

I implemented two other methods, so that the interpreter can get the note in the server, the actual running effect is not the best solution.

**1) By having the interpreter directly access the external storage of the note repository, such as HDFS, S3.**

1. But this limits the need to store notes in an external storage system.
2. In hdfs with kerberos authentication system enabled, not all users can access the note folder.

**2) Get the note from the server from the REST client in the interpreter.**

1. Need to let zeppelin/api/notes intercept in shrio.ini.
2. Or let the interpreter access the REST interface of the server through the user and password, but it is not safe.

In addition, in ZEPPELIN-4018 Workflow and orchestration, it is also necessary to query all the notes and paragraphs of the user in the workflow paragraph for the execution of the note.

[Improvement]

* [ ] - No task

* https://issues.apache.org/jira/browse/ZEPPELIN-4019

CI pass

* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No