[ZEPPELIN-4030] spark-scala-2.10 won't be built under some profile
### What is this PR for?
Before this PR, module spark-scala-2.10 is only built under spark profile which support both scala 2.10 and 2.11. But it should not be depend on which version of spark to build with. The spark profile should only affect the embedded spark distribution of spark interpreter, it should support multiple spark version no matter which profile it is built against.

### What type of PR is it?
[Bug Fix]

### Todos
* [ ] - Task

### What is the Jira issue?
* https://jira.apache.org/jira/browse/ZEPPELIN-4030

### How should this be tested?
* CI pass

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Jeff Zhang <zjffdu@apache.org>

Closes #3325 from zjffdu/ZEPPELIN-4030 and squashes the following commits:

aec33d24c [Jeff Zhang] [ZEPPELIN-4030] spark-scala-2.10 won't be built under some profile