[ZEPPELIN-4039] Fix restart interpreter process ConcurrentModificationException
1) I set it in zeppeln-env.sh
Export ZEPPELIN_FORCE_STOP=true
2) I create an interpreter process container in the remote YARN.
3) When I restart this interpreter, The interpreter process container will have a ConcurrentModificationException and will not exit properly.

The reason for starting this BUG is:
When exporting ZEPPELIN_FORCE_STOP=true,
Restart the interpreter process, the zeppelin server will execute the session that closes the process.
At the same time, the shutdown() and process() will be remote call.

In the RemoteInterpreterServer, the shutdown() and close() functions do not have concurrency control on the interpreterGroup. Causes an exception.