Pyspark needed a similar change
Now that we removed the isSecretSupported from the SparkVersion code we
need a way to determine whether the auth token can be used in Python.
Unlike R this uses Py4J, so if Py4J on the classpath has the ability to
use the AuthToken then we use the secret.