[ZEPPELIN-4216]. Introduce ZEPPELIN_SERVER_IP to specify the ip address of thrift server
### What is this PR for?

This PR use introduce env `ZEPPELIN_SERVER_IP` to specify the ip address of thrift server. It is used when you are in a complex network environment where yo might have multiple network cards or you use proxy. If it is not specified, it would still use the old approach to choose one IP randomly.

### What type of PR is it?
[ Improvement ]

### Todos
* [ ] - Task

### What is the Jira issue?
* https://jira.apache.org/jira/browse/ZEPPELIN-4216

### How should this be tested?
* CI pass

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Jeff Zhang <zjffdu@apache.org>

Closes #3392 from zjffdu/ZEPPELIN-4216 and squashes the following commits:

35fbcd51f [Jeff Zhang] [ZEPPELIN-4216]. Introduce ZEPPELIN_SERVER_IP to specify the ip address of thrift server