[ZEPPELIN-4219] User can start interpreter when interpreter dependencies jars are not downloaded
### What is this PR for?
Now interpreter can start even when it is downloading dependencies. It will cause ClassNotFound exception. The root cause is that the status of InterpreterSetting is incorrectly set to `READY`.  This PR fix this issue.

### What type of PR is it?
[Bug Fix ]

### Todos
* [ ] - Task

### What is the Jira issue?
* https://jira.apache.org/jira/browse/ZEPPELIN-4219

### How should this be tested?
* CI pass

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Jeff Zhang <zjffdu@apache.org>

Closes #3398 from zjffdu/ZEPPELIN-4219 and squashes the following commits:

d8d8309c6 [Jeff Zhang] [ZEPPELIN-4219] User can start interpreter when interpreter dependencies jars are not downloaded