[ZEPPELIN-4280] Fix an issue that makes an infinite loop while switching the visualizations
### What is this PR for?
Fix an issue that makes an infinite loop while switching the visualizations.

An infinite loop occurs because the front-end web application tries to commit the paragraph even if there are no changes in the visualization configuration.

### What type of PR is it?
Bug Fix

### What is the Jira issue?
https://issues.apache.org/jira/browse/ZEPPELIN-4280

### How should this be tested?
If you switch quickly between different visualizations multiple times, the infinite loops should not happen.

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Jaewon Seo <jaewon@zepl.com>

Closes #3414 from Byeol/ZEPPELIN-4280 and squashes the following commits:

d49d36ac1 [Jaewon Seo] Commit the paragraph only if there are changes in the visualization configuration