[ZEPPELIN-2680] Allow opening notebook as a reader
What is this PR for?
--------------------

When users only have reader permissions over a notebook AND the notebook
contains a chart, the app sends a COMMIT_PARAGRAPH for each paragraph,
which causes a write permission popup to show
("Insufficient privileges to write note").

Ths issue is with function commitParagraphResult in result.controller.js.

The original [fix](https://github.com/apache/zeppelin/pull/2439)
introduced a nasty bug that dramatically increased CPU usage and was never
merged.

This PR fixes the bug. It also automatically switches to `report` mode
as it is pointless to show UI features that the user is not allowed
to control.

What type of PR is it?
----------------------
Bug Fix

What is the Jira issue?
-----------------------
[ZEPPELIN-2680](https://issues.apache.org/jira/browse/ZEPPELIN-2680)

Questions:

- Does the licenses files need update? No
- Is there breaking changes for older versions? No
- Does this needs documentation? No