ZEPPELIN-4329 fix building Scio with Scala 2.11
Now both Scio & Beam interpreters explicitly use Scala 2.10, so it's possible
to compile both if Scala 2.11 profile is enabled.

Also was need to add missing dependency for Beam interpreter, otherwise it didn't work.

Plus small rework of the properties, so the Scala 2.10 version could be explicitly used in
the interpreters. (Also, tried Scala 2.10.7, and it didn't work, so we must stay on 2.10.5).