[ZEPPELIN-4263] Fixed cluster docker mode cannot exit the container of the remote interpreter
### What is this PR for?
Because docker first used the `while :; do sleep 1; done` script to start the container.

Execute `bin/interpreter.sh` to execute the interpreter startup script in the container.

When zeppelin closes or restarts the interpreter in docker, the interpreter process in the container can exit normally, but the container's startup shell does not exit, causing the container to continue running.

So we need to determine if the interpreter process exits in the container's startup script.

### What type of PR is it?
Bug Fix

### What is the Jira issue?
https://issues.apache.org/jira/browse/ZEPPELIN-4263

### How should this be tested?
[CI Pass](https://travis-ci.org/liuxunorg/zeppelin/builds/585091106)

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Xun Liu <liuxun@apache.org>

Closes #3445 from liuxunorg/ZEPPELIN-4263 and squashes the following commits:

017e98b01 [Xun Liu] 1. add comment. 2. add determine if container exists.
93e41e5fb [Xun Liu] [ZEPPELIN-4263] Fixed cluster docker mode cannot exit the container of the remote interpreter