[ZEPPELIN-4386] Resource.invokeMethod() doesn't find the right method to call
### What is this PR for?
The logic in Resource.invokeMethod() seems to be failing to find the correct method signature, if the method to invoke is overloaded and the parameter types for invokeMethod() are not specified.

### What type of PR is it?
Bug Fix

### Todos
--

### What is the Jira issue?
https://issues.apache.org/jira/browse/ZEPPELIN-4386

### How should this be tested?
New unit tests added to cover these scenarios.
[CI link](https://travis-ci.org/gkomlossi/zeppelin/builds/602501345)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Gyula Komlossi <gkomlossi@cloudera.com>

Closes #3492 from gkomlossi/master and squashes the following commits:

c4f919f1a [Gyula Komlossi] [ZEPPELIN-4386] Resource.invokeMethod() doesn't find the right method to call