[ZEPPELIN-4444] Fix schema disagreement when execute DDL statements
### What is this PR for?

When executing DDL statements (`CREATE`/`ALTER`/`DROP`) it's possible to get schema disagreement in Cassandra, especially for big/geo-distributed clusters. Java driver has a special handling of such statements, with increased timeout, but it's still possible that next DDL statement will be executed before all nodes agree on the schema version, and this could lead to schema disagreement that need to be resolved by administrators.

### What type of PR is it?
Improvement

### What is the Jira issue?

https://issues.apache.org/jira/browse/ZEPPELIN-4444

### How should this be tested?

* Added the unit test
* Tested manually
* Travis CI: https://travis-ci.org/alexott/zeppelin/builds/620614700

Author: Alex Ott <alexott@gmail.com>

Closes #3527 from alexott/ZEPPELIN-4444 and squashes the following commits:

4b803b7ab [Alex Ott] [ZEPPELIN-4444] Fix schema disagreement when execute DDL statements