[ZEPPELIN-4469]. Cron is disabled after zeppelin restart
### What is this PR for?
The root cause is that zeppelin didn't init cron job setting after zeppelin restart. This PR add one thread to list all notes to init the cron job. The reason to do in a separated thread is that there may be many notes, loop all notes in the main thread may block the restarting of Zeppelin server.

### What type of PR is it?
[Bug Fix]

### Todos
* [ ] - Task

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-4469

### How should this be tested?
* Test it manually.

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Jeff Zhang <zjffdu@apache.org>

Closes #3542 from zjffdu/ZEPPELIN-4469 and squashes the following commits:

7bc6491fd [Jeff Zhang] address comment
4b9876c5d [Jeff Zhang] [ZEPPELIN-4469]. Cron is disabled after zeppelin restart