[ZEPPELIN-4515] Have Jetty Server support multiple wars
### What is this PR for?
We have a new UI with version angular 8, which will be retiring the old angularjs.
both of the angular frameworks may need to co-exist for a while until the new UI is stable.
Have the Jetty server support multiple wars with different port (8080 and 9090)

### What type of PR is it?
[Feature ]

### Todos
retire angularJS completely in the future.

### What is the Jira issue?
* [ZEPPELIN-4515](https://issues.apache.org/jira/browse/ZEPPELIN-4515) Have Jetty Server support multiple wars
Add multiple war support for the Zeppelin Server.
two ports (default is 8080 and 9090) will be living for a while until we are comfortable with the new UI.

### How should this be tested?
The default port for new UI is 9090, testing via http://localhost:9090

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? **No**
* Is there breaking changes for older versions? **No**
* Does this needs documentation? **No**

Author: fdai <fdai@twitter.com>

Closes #3571 from fred521/add_angular_to_jetty and squashes the following commits:

196547605 [fdai] add next context path
93e2a4fdc [fdai] clean the code
8606ce064 [fdai] use one port two support multiple wars in jetty
f08602e20 [fdai] remove unrelated notebook
0f9c10367 [fdai] add angular support for jetty