[ZEPPELIN-4541]. Add api to checkpoint paragraph output between interpreter and zeppelin-server
### What is this PR for?
In ZEPPELIN-4525 we add support of Shiny app in R interpreter, because when we refresh the note page, the iframe is not displayed, the root cause is that we only save note when paragraph is finished, while we launch shiny app in zeppelin, the paragraph is in RUNNING state. In this PR, we add api to checkpoint paragraph output so that user can see the output even after refresh note.

### What type of PR is it?
[ Improvement ]

### Todos
* [ ] - Task

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-4541

### How should this be tested?
* CI pass

### Screenshots (if appropriate)
![shiny_app](https://user-images.githubusercontent.com/164491/71878752-6428a800-3167-11ea-92b5-f388650d19f5.gif)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Jeff Zhang <zjffdu@apache.org>

Closes #3589 from zjffdu/ZEPPELIN-4541 and squashes the following commits:

38f8ba8ba [Jeff Zhang] [ZEPPELIN-4541]. Add api to checkpoint paragraph output between interpreter and zeppelin-server