[ZEPPELIN-4573] Disable zeppelin-web.war extraction when pointing to a .war file
### What is this PR for?
This PR disables the local extraction of zeppelin-web.war if ZEPPELIN_WAR is pointing to a .war file. This has 2 advantages:
1. Skipping WAR extraction improves Zeppelin launch time
2. It prevents any problems arising from the corruption of local zeppelin-web extraction

### What type of PR is it?
Improvement

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-4573

THIS SOFTWARE IS CONTRIBUTED SUBJECT TO THE TERMS OF THE APACHE SOFTWARE FOUNDATION SOFTWARE GRANT AND CORPORATE CONTRIBUTOR LICENSE AGREEMENT VERSION R190612.

THIS SOFTWARE IS LICENSED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE AND ANY WARRANTY OF NON-INFRINGEMENT, ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE. THIS SOFTWARE MAY BE REDISTRIBUTED TO OTHERS ONLY BY EFFECTIVELY USING THIS OR ANOTHER EQUIVALENT DISCLAIMER IN ADDITION TO ANY OTHER REQUIRED LICENSE TERMS.

Author: Muhammad-ms <59560973+Muhammad-ms@users.noreply.github.com>
Author: Muhammad Taufiq <Muhammad.Taufiq@morganstanley.com>

Closes #3614 from Muhammad-ms/zeppelin9_war_extraction and squashes the following commits:

d01bdd2da [Muhammad Taufiq] [ZEPPELIN-4564] Disable zeppelin-web.war extraction when pointing to a .war file
4f532aad5 [Muhammad-ms] Merge pull request #1 from apache/master