[ZEPPELIN-4574] Fixed rootNotebookFolder and noteFileName in VFSNotebookRepo.java to allow it to work on windows
### What is this PR for?
The rootNotebookFolder variable in VFSNotebookRepo.java does not contain the root directory in windows. For e.g. if notebook directory is "C:/Users/zeppelin-notes", the value of rootNotebookFolder set in VFSNotebookRepo.java:91 is "/Users/zeppelin-notes", which is not found in windows and throws an exception at launch.
Using `this.rootNotebookFolder = rootNotebookFileObject.getName().getURI().replace("file:///", "/");` instead allows us to set the correct value for rootNotebookFolder (i.e. /C:/Users/zeppelin-notes).

Similarly, the noteFileName (VFSNotebookRepo.java:110) variable is missing root directory in windows and using `String noteFileName = fileObject.getName().getURI().replace("file:///", "/");` fixes this.

### What type of PR is it?
Bug Fix

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-4574

THIS SOFTWARE IS CONTRIBUTED SUBJECT TO THE TERMS OF THE APACHE SOFTWARE FOUNDATION SOFTWARE GRANT AND CORPORATE CONTRIBUTOR LICENSE AGREEMENT VERSION R190612.

THIS SOFTWARE IS LICENSED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE AND ANY WARRANTY OF NON-INFRINGEMENT, ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE. THIS SOFTWARE MAY BE REDISTRIBUTED TO OTHERS ONLY BY EFFECTIVELY USING THIS OR ANOTHER EQUIVALENT DISCLAIMER IN ADDITION TO ANY OTHER REQUIRED LICENSE TERMS.

Author: Muhammad Taufiq <Muhammad.Taufiq@morganstanley.com>
Author: Muhammad-ms <59560973+Muhammad-ms@users.noreply.github.com>

Closes #3615 from Muhammad-ms/zeppelin9_vfs and squashes the following commits:

b0a65c0d2 [Muhammad Taufiq] [ZEPPELIN-4574] Added comments explaining the change
81b054e92 [Muhammad Taufiq] [ZEPPELIN-4574] Fixed rootNotebookFolder and noteFileName in VFSNotebookRepo.java to allow it to work on windows
4f532aad5 [Muhammad-ms] Merge pull request #1 from apache/master