[ZEPPELIN-4615]. Remove annoying exception when shutting down interpreter
### What is this PR for?

The reason we see the annoying exception is because we call shutdown rpc in a blocking way, so before the rpc call return, the interpreter process is terminated. This PR would just wrap the shutdown in one thread in interpreter process and would return after the shutdown thread is started.

### What type of PR is it?
[Improvement]

### Todos
* [ ] - Task

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-4615

### How should this be tested?
* Tested manually.

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Jeff Zhang <zjffdu@apache.org>

Closes #3645 from zjffdu/ZEPPELIN-4615 and squashes the following commits:

ac6742e34 [Jeff Zhang] [ZEPPELIN-4615]. Remove annoying exception when shutting down interpreter