[ZEPPELIN-4661]. Fail to run whole note
### What is this PR for?

This is to fix the regression issue caused by ZEPPELIN-4661. The root cause is that in the frontend we use `paragraph` to represent the script text while in the backend we use `text`, so that converting between frontend data to backend class `Paragraph` will cause this kind of error.

### What type of PR is it?
[Bug Fix ]

### Todos
* [ ] - Task

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-4661

### How should this be tested?
* CI pass and manually tested

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? NO
* Is there breaking changes for older versions? No
* Does this needs documentation? NO

Author: Jeff Zhang <zjffdu@apache.org>

Closes #3672 from zjffdu/ZEPPELIN-4661 and squashes the following commits:

c0e6446a3 [Jeff Zhang] [ZEPPELIN-4661]. Fail to run whole note