[ZEPPELIN-4727] Fix HDFS Credentials storage
### What is this PR for?
With this PR we are switching to the abstract class `ConfigStorage` to save/load Credentials. We have child classes of `ConfigStorage` which are able to save to disc or HDFS.

NOTE: Cluster communication with changed user credentials is very simple. The `credential.json` content change is the only communication between cluster members. This leads to many rereads.

### What type of PR is it?
Bug Fix

### Todos
* [ ] - Task

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-4727

### How should this be tested?
* **Travis-CI**: https://travis-ci.org/github/Reamer/zeppelin/builds/673728600

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Philipp Dallig <philipp.dallig@gmail.com>

Closes #3718 from Reamer/credentials_storage and squashes the following commits:

6c8dd45e5 [Philipp Dallig] Use ConfigStage in Credentials and handle IOException in CredentialRestApi